Add option to build with V8 for javascript support
XFA is also enabled.

In order to make it easier for a consuming application a new method
FPDF_InitEmbeddedLibraries(char* resPath) is added.
This initializes v8 and other required libraries without the need
to include any v8 headers (or assume v8 knowledge on the users side)
when using the binaries.

A -v8 is appended to the archive filename to differentiate from the
default build option.