macOS: allow install_name_tool -id to work
This is required so that LC_ID_DYLIB may be set using install_name_tool
by downstream consumers. What the LC_ID_DYLIB field does is that when
the dylib is given to the linker, it forms the path the dylib in
question will be referred to in the RPATH of the resulting dependent
dylibs/executables.

This is necessary to support nixpkgs which wants to set this field to
/nix/store/..., but hits an error if we try doing that due to lack of
padding in the executable header (requiring relinking upstream, which,
hi!).

```
       > /nix/store/9n1n53pyrjdanm5gsxcgchr9v0i0401v-pdfium-lib/lib/libpdfium.dylib: fixing dylib
       > error: install_name_tool: changing install names or rpaths can't be redone for: /nix/store/9n1n53pyrjdanm5gsxcgchr9v0i0401v-p
dfium-lib/lib/libpdfium.dylib (for architecture arm64) because larger updated load commands do not fit (the program must be relinked, 
and you may need to use -headerpad or -headerpad_max_install_names)
```